# affiliate/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Paparazzi Affiliate Commerce Platform — v1 API
  version: 1.0.0
  description: |
    Public v1 API for the Paparazzi Affiliate Commerce Platform (looks, offers,
    links, integrations, earnings, payouts, programmes, publishers, disputes,
    contracts, suspense).

    This spec is generated from the real route handlers in
    `packages/api/src/routes/*.ts` — it documents only endpoints the code
    registers. There is no `GET /v1/links` (links are minted, then resolved
    by the redirect service at `GET /r/{token}`).

    ## Authentication (temporary stub)

    All `/v1/*` endpoints require a Bearer JWT (`Authorization: Bearer <token>`).
    Expected claims: `sub`, `org_id`, `role`. **The JWT verification in
    `src/middleware.ts` (`requireAuth`) is a temporary stub: it trusts the
    token's claims verbatim and must be replaced by a production identity
    provider** with membership validation (`memberships` table), issuer /
    audience / expiry policy. `GET /healthz` is the only unauthenticated route.

    ## Envelopes

    - Success: `{ "data": <payload>, "request_id": "<uuid>" }` (`ok()` helper).
    - Error: `{ "error": { "code": "<ERROR_CODE>", "message": "..." }, "request_id": "<uuid>" }`.
    - Every response carries an `X-Request-Id` header echoing `request_id`.
    - Error `code` is always one of the `ErrorCode` enum values
      (`packages/shared/src/errors.ts`); clients should branch on `code`,
      never on message text.

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        TEMPORARY STUB — `requireAuth` in `src/middleware.ts` verifies the JWT
        signature against `JWT_SECRET` and trusts the `sub`, `org_id`, `role`
        claims verbatim. Replace with a production IdP (membership validation,
        issuer/audience/expiry policy) before pilot.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 200
      description: |
        Optional idempotency key. First response per (key, org) is stored and
        replayed with `X-Idempotent-Replay: true`; 5xx is never stored.
    PathId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  headers:
    X-Request-Id:
      description: Echoes the envelope `request_id` for log correlation.
      schema:
        type: string
        format: uuid
    X-Idempotent-Replay:
      description: Present (`true`) when the body is a replay of a stored idempotent response.
      schema:
        type: string
        enum: ['true']
  schemas:
    ErrorCode:
      type: string
      enum:
        - PROGRAMME_NOT_APPROVED
        - OFFER_STALE
        - PROPERTY_FORBIDDEN
        - PUBLISHER_NOT_ACTIVE
        - PROPERTY_NOT_OWNER_OPERATED
        - NOT_FOUND
        - VALIDATION_ERROR
        - UNAUTHORIZED
        - FORBIDDEN
        - CONFLICT
        - RATE_LIMITED
        - UPSTREAM_UNAVAILABLE
        - LEDGER_IMBALANCE
        - TRANSFER_STATUS_UNKNOWN
        - INTERNAL
      description: Canonical codes from `packages/shared/src/errors.ts`.
    ErrorEnvelope:
      type: object
      required: [error, request_id]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
        request_id:
          type: string
          format: uuid
    SuccessEnvelope:
      type: object
      required: [data, request_id]
      properties:
        data: {}
        request_id:
          type: string
          format: uuid
      description: Shape of every 2xx body; `data` is operation-specific.
    Uuid:
      type: string
      format: uuid
...
    DateTime:
      type: string
      format: date-time
...
    DisputeKind:
      type: string
      enum: [missing_commission, wrong_amount, unattributed_click, other]
    DisputeStatus:
      type: string
      enum: [open, under_review, resolved, rejected]
    Dispute:
      type: object
      required: [id, kind, subject, status]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        conversion_id: { $ref: '#/components/schemas/Uuid', nullable: true }
        publisher_id: { $ref: '#/components/schemas/Uuid', nullable: true }
        kind: { $ref: '#/components/schemas/DisputeKind' }
        subject: { type: string }
        claim_ref: { type: string, nullable: true, description: e.g. the merchant's order id, for missing_commission. }
        status: { $ref: '#/components/schemas/DisputeStatus' }
        evidence: { type: object, nullable: true, description: Free-form evidence object. Screenshots alone never create a payable sale. }
        resolution_note: { type: string, nullable: true }
        resolved_by: { type: string, nullable: true }
        resolved_at: { $ref: '#/components/schemas/DateTime', nullable: true }
        created_at: { $ref: '#/components/schemas/DateTime' }
      description: A ticket NEVER creates a payable sale on its own.
...
  responses:
    BadRequest:
      description: '400 — request failed validation (VALIDATION_ERROR).'
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: '401 — missing/invalid Bearer token or claims (UNAUTHORIZED).'
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: '403 — role not allowed (FORBIDDEN), or a link-guard rejection (PROPERTY_FORBIDDEN, PROGRAMME_NOT_APPROVED, PUBLISHER_NOT_ACTIVE, PROPERTY_NOT_OWNER_OPERATED).'
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: '404 — no route, unknown connector, or resource not in this org (NOT_FOUND).'
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    InternalError:
      description: '500 — unexpected failure (INTERNAL); message not leaked.'
      headers:
        X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
paths:
...
  /v1/disputes:
    post:
      tags: [Disputes]
      summary: Open a dispute ticket
      description: |
        Opens a ticket for a missing commission or related claim. It may
        reference a known conversion (`wrong_amount` / `unattributed_click`) or
        NO conversion at all (`missing_commission`: the publisher believes a
        sale happened that the provider never reported — identified by
        `claim_ref`, e.g. the merchant's order id). When a conversion is
        named it must belong to this org (`404` otherwise); `publisher_id` is
        derived from the conversion's click chain unless explicitly given.
        A ticket NEVER creates a payable sale on its own.
        Roles: `publisher_owner`, `editor`, `network_admin`.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subject]
              properties:
                conversion_id: { $ref: '#/components/schemas/Uuid' }
                publisher_id: { $ref: '#/components/schemas/Uuid' }
                kind: { $ref: '#/components/schemas/DisputeKind', default: missing_commission }
                subject: { type: string, minLength: 1, maxLength: 300, example: Commission missing for order ORD-88412 }
                claim_ref: { type: string, minLength: 1, maxLength: 200, example: ORD-88412 }
                evidence: { type: object, description: Free-form evidence object. }
            examples:
              missing:
                value:
                  kind: missing_commission
                  subject: Commission missing for order ORD-88412
                  claim_ref: ORD-88412
                  evidence: { order_screenshot: attached }
      responses:
        '201':
          description: Ticket opened (status `open`); emits `dispute.opened`.
          headers:
            X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
            X-Idempotent-Replay: { $ref: '#/components/headers/X-Idempotent-Replay' }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    required: [data]
                    properties:
                      data: { $ref: '#/components/schemas/Dispute' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    get:
      tags: [Disputes]
      summary: List disputes
      description: Tenant-scoped list, newest first, with optional filters. Any authenticated role.
      parameters:
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/DisputeStatus' }
        - name: kind
          in: query
          schema: { $ref: '#/components/schemas/DisputeKind' }
        - name: publisher_id
          in: query
          schema: { $ref: '#/components/schemas/Uuid' }
      responses:
        '200':
          description: Disputes.
          headers:
            X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    required: [data]
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Dispute' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /v1/disputes/{id}:
    get:
      tags: [Disputes]
      summary: Get a dispute
      description: Tenant-scoped read. Any authenticated role.
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: The dispute.
          headers:
            X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    required: [data]
                    properties:
                      data: { $ref: '#/components/schemas/Dispute' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/disputes/{id}/resolve:
    post:
      tags: [Disputes]
      summary: Resolve a dispute
      description: |
        Outcome `resolved` requires `provider_verified: true` plus a resolution
        note describing the provider-side evidence — "screenshots alone never
        create a payable sale" is enforced in code (`422`), not just policy.
        Resolving a missing-commission ticket does NOT post ledger entries:
        the provider must re-report the conversion through the normal webhook
        path (dedupe + revision ordering apply), or finance books a manual
        correction. Emits `dispute.resolved` and writes `audit_log`.
        Roles: `network_admin`, `editor`.
      parameters:
        - $ref: '#/components/parameters/PathId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [outcome, resolution_note, provider_verified]
              properties:
                outcome: { type: string, enum: [resolved, rejected] }
                resolution_note: { type: string, minLength: 10, maxLength: 2000 }
                provider_verified: { type: boolean, description: Must be true for outcome=resolved — the claim was checked against provider records. }
            examples:
              resolve:
                value:
                  outcome: resolved
                  resolution_note: Provider confirmed txn-0001 at revision 2; commission re-reported via webhook.
                  provider_verified: true
      responses:
        '200':
          description: Dispute resolved/rejected.
          headers:
            X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
            X-Idempotent-Replay: { $ref: '#/components/headers/X-Idempotent-Replay' }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    required: [data]
                    properties:
                      data: { $ref: '#/components/schemas/Dispute' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: '`CONFLICT` — dispute is already resolved or rejected.'
          headers:
            X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '422':
          description: '`VALIDATION_ERROR` — outcome=resolved requires provider_verified=true.'
          headers:
            X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '500': { $ref: '#/components/responses/InternalError' }
...